[FEATURE] modify: changed attribute type

+++ src/main/java/com/spring/factorial/model/Calcs.java
@@ -6,15 +6,16 @@ import javax.persistence.*;
 @Entity
 @Table(name="TB_CALCS")
 public class Calcs {
-
+    // Atributos de classe
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Integer id;
 
-    private Integer number;
+    private Long number;
 
-    private Integer result;
+    private Long result;
 
+    // Metodos dos atributos
     public Integer getId(){
         return this.id;
     }
@@ -23,19 +24,19 @@ public class Calcs {
         this.id = id;
     }
 
-    public Integer getNumber(){
+    public Long getNumber(){
         return this.number;
     }
 
-    public void setNumber(Integer number){
+    public void setNumber(Long number){
         this.number = number;
     }
 
-    public Integer getResult(){
+    public Long getResult(){
         return this.result;
     }
 
-    public void setResult(Integer result){
+    public void setResult(Long result){
         this.result = result;
     }
